@@ -1,23 +1,23 @@
 package com.livinglifetechway.k4kotlinsample
 
-import android.databinding.DataBindingUtil
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
-import com.livinglifetechway.k4kotlin.fadeIn
-import com.livinglifetechway.k4kotlin.fadeOut
-import com.livinglifetechway.k4kotlin.hide
-import com.livinglifetechway.k4kotlin.hideKeyboard
-import com.livinglifetechway.k4kotlin.hideViews
-import com.livinglifetechway.k4kotlin.invisible
-import com.livinglifetechway.k4kotlin.onItemSelected
-import com.livinglifetechway.k4kotlin.orZero
-import com.livinglifetechway.k4kotlin.setItems
-import com.livinglifetechway.k4kotlin.show
-import com.livinglifetechway.k4kotlin.showViews
-import com.livinglifetechway.k4kotlin.toggleVisibility
+import androidx.appcompat.app.AppCompatActivity
+import androidx.databinding.DataBindingUtil
+import com.livinglifetechway.k4kotlin.core.fadeIn
+import com.livinglifetechway.k4kotlin.core.fadeOut
+import com.livinglifetechway.k4kotlin.core.hide
+import com.livinglifetechway.k4kotlin.core.hideKeyboard
+import com.livinglifetechway.k4kotlin.core.hideViews
+import com.livinglifetechway.k4kotlin.core.invisible
+import com.livinglifetechway.k4kotlin.core.onItemSelected
+import com.livinglifetechway.k4kotlin.core.orZero
+import com.livinglifetechway.k4kotlin.core.setItems
+import com.livinglifetechway.k4kotlin.core.show
+import com.livinglifetechway.k4kotlin.core.showViews
+import com.livinglifetechway.k4kotlin.core.startActivity
+import com.livinglifetechway.k4kotlin.core.toast
+import com.livinglifetechway.k4kotlin.core.toggleVisibility
 import com.livinglifetechway.k4kotlinsample.databinding.ActivityMainBinding
-import org.jetbrains.anko.startActivity
-import org.jetbrains.anko.toast
 
 class MainActivity : AppCompatActivity() {
 
